pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: BSD License",
]
name = "nomad-inl-base"
description = "A plugin to manage the data from LaNaSC"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Diego Garzon", email = "diego.garzon@inl.int" },
]
maintainers = [
    { name = "Diego Garzon", email = "diego.garzon@inl.int" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.4.1",
    "python-magic-bin; sys_platform == 'win32'",
    'nomad-material-processing',
    "nomad-baseclasses @ git+https://github.com/nomad-hzb/nomad-baseclasses.git@a569524f06e7f9fc4996fee2bb9d54b88930f4e1",
    "pdfplumber",
    "nomad-analysis",
]

[project.urls]
Repository = "https://github.com/GarzonDiegoFEUP/nomad-inl-base"

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest",
    "structlog",
    "mkdocs",
    "mkdocs-material==8.1.1",
    "pymdown-extensions",
    "mkdocs-click",
    "pytest-asyncio",
]

[tool.uv]
extra-index-url = [
    "https://gitlab.mpcdf.mpg.de/api/v4/projects/2187/packages/pypi/simple",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
    "F405", # May be undefined, or defined from star imports
    "PLC0415", # `import` should be at the top-level of a file
    "E501", # Line too long
    "PLR0912", # Too many branches
    "PLR0913", # Too many arguments in function definition
    "PLR0915", # Too many statements
    "PLR2004", # Magic value used in comparison
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["PLC0415"] # import-outside-top-level

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
CVarser_entry_point = "nomad_inl_base.parsers:CV_parser_entry_point"
EDparser_entry_point1 = "nomad_inl_base.parsers:ED_parser_entry_point"
schema_package_entry_point = "nomad_inl_base.schema_packages:schema_package_entry_point"
star_entry_point = "nomad_inl_base.schema_packages:star_entry_point"
crystaLLM_entry_point = "nomad_inl_base.schema_packages:crystaLLM_entry_point"
wet_deposition_entry_point = "nomad_inl_base.schema_packages:wet_deposition_entry_point"
characterization_entry_point = "nomad_inl_base.schema_packages:characterization_entry_point"
entities_entry_point = "nomad_inl_base.schema_packages:entities_entry_point"
cleaning_entry_point = "nomad_inl_base.schema_packages:cleaning_entry_point"
batteries_entry_point = "nomad_inl_base.schema_packages:batteries_entry_point"
annealing_entry_point = "nomad_inl_base.schema_packages:annealing_entry_point"
pc03_parser_entry_point = "nomad_inl_base.parsers:pc03_parser_entry_point"
pc04_parser_entry_point = "nomad_inl_base.parsers:pc04_parser_entry_point"
four_point_probe_parser_entry_point = "nomad_inl_base.parsers:four_point_probe_parser_entry_point"
kla_tencor_profiler_parser_entry_point = "nomad_inl_base.parsers:kla_tencor_profiler_parser_entry_point"
eqe_parser_entry_point = "nomad_inl_base.parsers:eqe_parser_entry_point"
solar_cell_iv_parser_entry_point = "nomad_inl_base.parsers:solar_cell_iv_parser_entry_point"
gdoes_parser_entry_point = "nomad_inl_base.parsers:gdoes_parser_entry_point"
analysis_entry_point = "nomad_inl_base.analysis:analysis_entry_point"



[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]